--- src/components/common/audio/click-sound/ClickSound.module.css
@@ -1,17 +1,0 @@
-.clickButton {
-	cursor: pointer;
-	padding: 8px 16px;
-	border-radius: 4px;
-	border: 1px solid #ccc;
-	background-color: #f0f0f0;
-	transition: all 0.2s ease;
-	outline: none;
-}
-
-.clickButton:hover {
-	background-color: #e0e0e0;
-}
-
-.clickButton:active {
-	transform: scale(0.98);
-}
