+++ src/features/items/components/item-card-list-client/ItemCardListClient.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import Image from "next/image";
 import styles from "./ItemCardListClient.module.css";
 import { useRouter } from "next/navigation";
-import { useClickSound } from "@/components/common/audio/click-sound/ClickSound";
+import { useClickSound } from "@/hooks/useClickSound";
 import { Item } from "@/features/items/types/items.types";
 import { customItemSilhouetteImages } from "@/features/items/data/itemsData";
 import ItemListSkeleton from "@/features/items/components/item-list-skeleton/ItemListSkeleton";
