+++ src/features/logs/components/logs-page-footer/LogsPageFooter.tsx
@@ -4,7 +4,7 @@ import styles from "./LogsPageFooter.module.css";
 import Link from "next/link";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { useClickSound } from "@/components/common/audio/click-sound/ClickSound";
+import { useClickSound } from "@/hooks/useClickSound";
 
 const LogsPageFooter = () => {
 	const router = useRouter();
