+++ src/features/party/components/party-member-card-list-client/PartyMemberCardListClient.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import styles from "./PartyMemberCardListClient.module.css";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { useClickSound } from "@/components/common/audio/click-sound/ClickSound";
+import { useClickSound } from "@/hooks/useClickSound";
 import { PartyMember } from "@/features/party/types/party.types";
 import { customMemberSilhouetteImages } from "@/features/party/data/partyMemberData";
 import PartyListSkeleton from "@/features/party/components/party-list-skeleton/PartyListSkeleton";
